+++ examples/README.md
@@ -25,7 +25,7 @@ to verify behaviour across the SDK.
 | [`database_examples/`](database_examples/) | Smoke tests per supported backend | 5 |
 | [`demos/`](demos/) | Feature demos — broadest coverage | 22 |
 | [`guides/`](guides/) | Short focused how-to guides | 13 |
-| [`integrations/`](integrations/) | Integrations Hub — first-class SaaS connectors (Gmail, …) | 1 |
+| [`integrations/`](integrations/) | Integrations Hub — first-class SaaS connectors (Gmail, Notion, …) | 1 |
 | [`pocs/`](pocs/) | Research-grade proofs of concept (entity disambiguation, canonicalization, prefetch) | 7 |
 
 Most runnable demos and backend examples use the v1.0 memory API (`remember`, `recall`, `forget`, `improve`). The lower-level `add`, `cognify`, `search`, and `prune` calls are intentionally kept in examples that demonstrate pipeline internals, permissions, relational migrations, or research POCs.
@@ -91,6 +91,7 @@ Most runnable demos and backend examples use the v1.0 memory API (`remember`, `r
 | [`web_url_content_ingestion_example.py`](demos/web_url_content_ingestion_example.py) | Crawl a URL and cognify the content |
 | [`dlt_ingestion_example.py`](demos/dlt_ingestion_example.py) | Ingest via [dlt](https://dlthub.com/) sources |
 | [`gmail_connector_example.py`](demos/gmail_connector_example.py) | "Ask my inbox" — Gmail connector (incremental + forget-on-delete) |
+| [`notion_connector_example.py`](demos/notion_connector_example.py) | Notion pages → memory (incremental + forget-on-delete) |
 | [`multimedia_processing/multimedia_audio_image_processing_example.py`](demos/multimedia_processing/multimedia_audio_image_processing_example.py) | Audio + image ingestion |
 | [`simple_document_qa/simple_document_qa_demo.py`](demos/simple_document_qa/simple_document_qa_demo.py) | Q&A over a single document |
 | [`simple_relational_database_migration_example/simple_relational_database_migration_example.py`](demos/simple_relational_database_migration_example/simple_relational_database_migration_example.py) | SQL → graph (small schema) |
@@ -167,14 +168,12 @@ Same scripts, indexed by what they demonstrate.
 ### Connectors / Integrations Hub
 - [`integrations/`](integrations/) — first-class SaaS connectors (DLT-based)
 - [`demos/gmail_connector_example.py`](demos/gmail_connector_example.py) — Gmail ("ask my inbox")
+- [`demos/notion_connector_example.py`](demos/notion_connector_example.py) — Notion pages (incremental + forget-on-delete)
 
 ### SQL → knowledge graph
 - [`custom_pipelines/relational_database_to_knowledge_graph_migration_example.py`](custom_pipelines/relational_database_to_knowledge_graph_migration_example.py)
 - [`demos/simple_relational_database_migration_example/simple_relational_database_migration_example.py`](demos/simple_relational_database_migration_example/simple_relational_database_migration_example.py)
 - [`demos/complex_relational_database_migration_example/complex_relational_database_migration_example.py`](demos/complex_relational_database_migration_example/complex_relational_database_migration_example.py)
-
-### External connectors (dlt)
-- [`python/notion_connector_example.py`](python/notion_connector_example.py) — Notion pages with incremental sync + forget-on-delete (needs `cognee[notion]` and `NOTION_API_KEY`)
 
 ### Custom pipelines / tasks
 - [`custom_pipelines/custom_cognify_pipeline_example.py`](custom_pipelines/custom_cognify_pipeline_example.py)
